@@ -2,90 +2,54 @@
   "name": "lan-play-status",
   "version": "0.1.0",
   "license": "AGPL-3.0-or-later",
+  "type": "module",
   "scripts": {
-    "preserve": "node build-flags.js",
-    "serve": "vue-cli-service serve",
+    "predev": "node build-flags.js",
+    "dev": "vite",
     "prebuild": "node build-flags.js",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "test:e2e": "vue-cli-service test:e2e",
-    "lint": "vue-cli-service lint",
-    "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\""
+    "build": "vite build",
+    "preview": "vite preview",
+    "test:unit": "vitest",
+    "test:e2e": "playwright test",
+    "lint": "eslint .",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "core-js": "^3.6.5",
-    "register-service-worker": "^1.7.1",
-    "vue": "^3.0.0",
+    "@intlify/unplugin-vue-i18n": "^11.0.1",
+    "pinia": "^3.0.3",
+    "vue": "^3.5.22",
     "vue-gtag-next": "^1.14.0",
-    "vue-i18n": "^9.0.0-rc.7",
-    "vue-router": "^4.0.0-0",
-    "vuex": "^4.0.0-0"
+    "vue-i18n": "^11.1.12",
+    "vue-router": "^4.6.3"
   },
   "devDependencies": {
-    "@intlify/vue-i18n-loader": "^2.0.0-rc.1",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-e2e-cypress": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-pwa": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "^2.0.0-0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^7.0.0-0",
-    "flagpack-core": "^1.0.0",
-    "geoip-country": "^4.0.11",
+    "@eslint/js": "^9.37.0",
+    "@playwright/test": "^1.56.1",
+    "@vitejs/plugin-vue": "^6.0.1",
+    "@vitest/eslint-plugin": "^1.3.23",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "@vue/test-utils": "^2.4.6",
+    "eslint": "^9.37.0",
+    "eslint-plugin-playwright": "^2.2.2",
+    "eslint-plugin-vue": "~10.5.0",
+    "flagpack-core": "^2.0.0",
+    "geoip-country": "^5.0.202511112342",
+    "globals": "^16.4.0",
     "he": "^1.2.0",
-    "node-sass": "^4.12.0",
-    "prettier": "^1.19.1",
-    "sass-loader": "^8.0.2",
-    "sitemap-webpack-plugin": "^1.1.0",
-    "svgo": "^2.2.2",
-    "typescript": "~3.9.3",
-    "vue-cli-plugin-i18n": "~2.0.0",
-    "vue-jest": "^5.0.0-0",
-    "xml2js": "^0.5.0"
+    "jsdom": "^27.0.1",
+    "prettier": "3.6.2",
+    "sass": "^1.94.0",
+    "sass-loader": "^16.0.6",
+    "sitemap-webpack-plugin": "^1.1.1",
+    "svgo": "^4.0.0",
+    "vite": "^7.1.11",
+    "vite-plugin-vue-devtools": "^8.0.3",
+    "vitest": "^3.2.4",
+    "vue-cli-plugin-i18n": "~2.3.2",
+    "xml2js": "^0.6.2"
   },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended",
-      "@vue/prettier"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {},
-    "overrides": [
-      {
-        "files": [
-          "**/__tests__/*.{j,t}s?(x)",
-          "**/tests/unit/**/*.spec.{j,t}s?(x)"
-        ],
-        "env": {
-          "jest": true
-        }
-      }
-    ]
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ],
-  "jest": {
-    "preset": "@vue/cli-plugin-unit-jest",
-    "transform": {
-      "^.+\\.vue$": "vue-jest"
-    }
+  "packageManager": "pnpm@10.20.0",
+  "engines": {
+    "node": "22"
   }
 }
