@@ -1,4 +1,2 @@
-module.exports = {
-  ldnMitmVersion: "v0.2.3",
-  ldnMitmWin32Version: "v0.2.1"
-};
+export const ldnMitmVersion = 'v0.2.3'
+export const ldnMitmWin32Version = 'v0.2.1'
