@@ -1,32 +1,42 @@
-import { createI18n } from "vue-i18n";
+import { createI18n } from 'vue-i18n'
 
-/**
- * Load locale messages
- *
- * The loaded `JSON` locale messages is pre-compiled by `@intlify/vue-i18n-loader`, which is integrated into `vue-cli-plugin-i18n`.
- * See: https://github.com/intlify/vue-i18n-loader#rocket-i18n-resource-pre-compilation
- */
-function loadLocaleMessages() {
-  const locales = require.context(
-    "./locales",
-    true,
-    /[A-Za-z0-9-_,\s]+\.json$/i
-  );
-  const messages = {};
-  locales.keys().forEach(key => {
-    const matched = key.match(/([A-Za-z0-9-_]+)\./i);
-    if (matched && matched.length > 1) {
-      const locale = matched[1];
-      messages[locale] = locales(key).default;
-    }
-  });
-  return messages;
+const fallbackLocale = 'en'
+const locale = navigator.language || fallbackLocale
+
+const loadLocaleMessages = async (locale) => {
+  try {
+    const messages = await import(`./locales/${locale}.json`)
+    return messages.default
+  } catch (e) {
+    // no-op
+  }
 }
 
-export default createI18n({
+const loadLocale = async (locale) => {
+  const messages = await loadLocaleMessages(locale)
+
+  i18n.global.setLocaleMessage(locale, messages)
+}
+
+const loadAndSetLocale = async (locale) => {
+  await loadLocale(locale)
+
+  i18n.global.locale.value = locale
+}
+
+const i18n = createI18n({
   legacy: false,
-  locale: navigator.language || process.env.VUE_APP_I18N_LOCALE || "en",
-  fallbackLocale: process.env.VUE_APP_I18N_FALLBACK_LOCALE || "en",
-  messages: loadLocaleMessages(),
-  globalInjection: true
-});
+  locale,
+  fallbackLocale,
+  messages: {},
+  numberFormats: {},
+  datetimeFormats: {},
+})
+
+loadLocale(locale)
+
+if (locale !== fallbackLocale) {
+  loadLocale(fallbackLocale)
+}
+
+export { i18n, loadAndSetLocale }
