@@ -1,5 +1,5 @@
-const getFullAddress = _server => {
-  return `${_server.ip}:${_server.port}`;
-};
+const getFullAddress = (_server) => {
+  return `${_server.ip}:${_server.port}`
+}
 
-export { getFullAddress };
+export { getFullAddress }
