@@ -1,6 +1,6 @@
 <template>
   <div class="install">
-    <h1>{{ $t("pages.installSwitch.title") }}</h1>
+    <h1>{{ $t('pages.installSwitch.title') }}</h1>
     <InstallConfig
       type="Switch"
       owner="90DNS"
@@ -8,41 +8,29 @@
       secondary="207.246.121.77"
     />
     <p>
-      <img
-        alt="Settings"
-        class="screenshot"
-        :src="require(`@/assets/switch/settings1.jpg`)"
-      />
+      <img alt="Settings" class="screenshot" src="@/assets/switch/settings1.jpg" />
     </p>
     <p>
-      <img
-        alt="Settings"
-        class="screenshot"
-        :src="require(`@/assets/switch/settings2.jpg`)"
-      />
+      <img alt="Settings" class="screenshot" src="@/assets/switch/settings2.jpg" />
     </p>
     <p>
-      <img
-        alt="Settings"
-        class="screenshot"
-        :src="require(`@/assets/switch/settings3.jpg`)"
-      />
+      <img alt="Settings" class="screenshot" src="@/assets/switch/settings3.jpg" />
     </p>
-    <Install type="switch" />
-    <h2>{{ $t("pages.installSwitch.test.title", { prefix: "4 - " }) }}</h2>
+    <InstallSteps type="switch" />
+    <h2>{{ $t('pages.installSwitch.test.title', { prefix: '4 - ' }) }}</h2>
     <p>
-      If you believe you have followed the instructions so far, you can try to
-      play a LAN game, such as Mario Kart. Sometimes you may see output in the
-      client, like <strong>IConnection::IConnection</strong>, which indicates
-      your switch is connected. Not all system combinations will show this.
+      If you believe you have followed the instructions so far, you can try to play a LAN game, such
+      as Mario Kart. Sometimes you may see output in the client, like
+      <strong>IConnection::IConnection</strong>, which indicates your switch is connected. Not all
+      system combinations will show this.
     </p>
     <p>
       <strong>
-        Failing the internet and/or DNS connectivity test on your switch DOES
-        NOT MATTER and can be ignored.
+        Failing the internet and/or DNS connectivity test on your switch DOES NOT MATTER and can be
+        ignored.
       </strong>
     </p>
-    <Terminal
+    <TerminalText
       title="user@linux: ~/Downloads"
       content="[DEBUG]: IConnection::IConnection
 [DEBUG]: TcpConnection::onConnection
@@ -52,16 +40,14 @@
 [DEBUG]: IConnection::~IConnection"
     />
     <p>
-      If you cannot join or create lobbies after asking someone to play with
-      you, join #help on the lan-play discord for further assistance.
-      <strong
-        >Note: You must ask people to play with you or to join others.</strong
-      >
+      If you cannot join or create lobbies after asking someone to play with you, join #help on the
+      lan-play discord for further assistance.
+      <strong>Note: You must ask people to play with you or to join others.</strong>
     </p>
-    <h2>{{ $t("pages.installSwitch.module.title", { prefix: "5 - " }) }}</h2>
+    <h2>{{ $t('pages.installSwitch.module.title', { prefix: '5 - ' }) }}</h2>
     <p>
-      You will need to download ldn_mitm if you wish to play to play games that
-      don't have the lan play feature (Local Wireless mode Games).
+      You will need to download ldn_mitm if you wish to play to play games that don't have the lan
+      play feature (Local Wireless mode Games).
     </p>
     <p>
       You can download the latest ldn_mitm
@@ -74,8 +60,8 @@
       >.
     </p>
     <p>
-      If you blindly upgraded AMS without checking compatibility, you will need
-      to use untested/unsupported nightly builds of ldn_mitm, which are
+      If you blindly upgraded AMS without checking compatibility, you will need to use
+      untested/unsupported nightly builds of ldn_mitm, which are
       <a
         href="https://github.com/spacemeowx2/ldn_mitm/actions"
         target="_blank"
@@ -85,26 +71,26 @@
       >.
     </p>
     <p>
-      After you downloaded it extract to the root of your SD card and the module
-      should load on your next reboot.
+      After you downloaded it extract to the root of your SD card and the module should load on your
+      next reboot.
     </p>
     <p>
-      After that you just launch your game of choice and select local wireless
-      play and you should be good to go.
+      After that you just launch your game of choice and select local wireless play and you should
+      be good to go.
     </p>
   </div>
 </template>
 
 <script>
-import InstallConfig from "@/components/InstallConfig.vue";
-import Install from "@/components/Install.vue";
-import Terminal from "@/components/Terminal.vue";
+import InstallConfig from '@/components/InstallConfig.vue'
+import InstallSteps from '@/components/InstallSteps.vue'
+import TerminalText from '@/components/TerminalText.vue'
 
 export default {
   components: {
     InstallConfig,
-    Install,
-    Terminal
-  }
-};
+    InstallSteps,
+    TerminalText,
+  },
+}
 </script>
